+++ src/Core/src/Platform/iOS/MauiWKWebView.cs
@@ -182,33 +182,11 @@ namespace Microsoft.Maui.Platform
 
 		bool LoadFile(string url)
 		{
-			try
-			{
-				var file = Path.GetFileNameWithoutExtension(url);
-				var ext = Path.GetExtension(url);
-				var directory = Path.GetDirectoryName(url);
+			ILogger? logger = null;
+			if (_handler.TryGetTarget(out var handler))
+				logger = handler.MauiContext?.CreateLogger<MauiWKWebView>();
 
-				// If there's a subdirectory, use the overload that accepts a subdirectory parameter else fallback to the original method if subdirectory method fails or if no subdirectory
-				NSUrl? nsUrl = string.IsNullOrEmpty(directory)
-					? NSBundle.MainBundle.GetUrlForResource(file, ext)
-					: NSBundle.MainBundle.GetUrlForResource(file, ext, directory);
-
-				if (nsUrl is null)
-				{
-					return false;
-				}
-
-				LoadFileUrl(nsUrl, nsUrl);
-
-				return true;
-			}
-			catch (Exception ex)
-			{
-				if (_handler.TryGetTarget(out var handler))
-					handler.MauiContext?.CreateLogger<MauiWKWebView>()?.LogWarning($"Could not load {url} as local file: {ex}");
-			}
-
-			return false;
+			return this.LoadFile(url, logger);
 		}
 
 		[UnconditionalSuppressMessage("Memory", "MEM0002", Justification = IUIViewLifeCycleEvents.UnconditionalSuppressMessage)]
